feat: implement user management functionality; add user model, CRUD operations, and health check endpoint with PostgreSQL integration
